# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skillbytes API - PBL Module",
        "description": "REST API for Problem-Based Learning (PBL) Case Management System",
        "contact": {
            "email": "support@skillbytes.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development"
        },
        {
            "url": "http://145.79.13.180/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/my-results": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "Get all user's assessment results",
                "description": "Retrieve all completed assessment attempts for the authenticated user",
                "operationId": "f74bd66c5def295eec175f14f31ef918",
                "responses": {
                    "200": {
                        "description": "List of user's results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "user_id": {
                                                        "type": "string"
                                                    },
                                                    "assessment": {
                                                        "type": "object"
                                                    },
                                                    "score": {
                                                        "type": "number"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "started_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "completed_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/my-results/{attemptId}": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "Get user's specific result detail",
                "description": "Retrieve detailed result of a specific assessment attempt including all answers, options, and correct answers for review",
                "operationId": "d29cb4912b76b6359dfdedb071037b2e",
                "parameters": [
                    {
                        "name": "attemptId",
                        "in": "path",
                        "description": "Assessment Attempt ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assessment result detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Result not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}